update commit_index on leader
Pay extra attention to the math: we need a median _index_. We subtract
before dividing so that we get something like this:

    cluster size | 1 | 2 | 3 | 4 | 5 | 6 …
    -------------+---+---+---+---+---+----
     median idx  | 0 | 0 | 1 | 1 | 2 | 2 …
    -------------+---+---+---+---+---+----
       above     | 1 | 2 | 2 | 3 | 3 | 4 …

Without this subtraction, we can overcommit (a 2-cluster followers the
leader, all even clusters follow a slim minority). See how above = size
- idx tallies the _number_ of servers at or above the median, and it's a
majority.

When we do the math wrong, there are too few servers at or above the
median, and we commit without a majority.

    cluster size | 1 | 2 | 3 | 4 | 5 | 6 …
    -------------+---+---+---+---+---+----
      bad   idx  | 0 | 1 | 2 | 2 | 3 | 3 …
    -------------+---+---+---+---+---+----
       above     | 1 | 1 | 1 | 2 | 2 | 3 …