make each node track the cluster size
This should allow me to unwind hacks in the network from the previous
commit, but comes at a slight cost:

- Construction/deserialization requires more info.
- The receive_vote function gets some version of clearer or murkier,
  depending on your perspective.

Let's pull the latter apart: it's clearer because the code reads like
you'd expect, but if you're thinking about the type you might be
confused: shouldn't this only happen if we're a candidate? Yes! The new
helpers do the right thing for candidates, so it's OK.

Why do we have to do this? Well, under "match &mut self.t" we can't get
back to an "&self" reference we can call "has_majority" on with
"voters.len()" (because it's using the mutable borrow of self). We might
be able to do "let v = voters.len()" first and have it all work, but
that seemed fragile, and I like this code split better.

One other advantage: we're able to more accurately track commit indices
in the tests this way.