test degenerate election in a 1-cluster
Notably, this exposed a major bug! Candidates weren't counting their
self-vote in the majority calculation, so would have waited longer for
more votes than necessary. For example, in a 3-cluster, we would wait
for both "other" votes instead of just one.

The Output changes for conversion to a candidate, dropping some outgoing
requests… but it shouldn't happen in bigger clusters (assert it is so).

Best-viewed-with: --ignore-space-change