fix commit_index bug
When a new leader is elected, their match indices will be all 0 as Raft
specifies. But this means that after receiving the first
AppendEntriesResponse, the median will probably be 0! We panic when see
we're trying to move the commit_index backward (good), but we haven't
implemented Raft (bad).

Instead, follow the spec: increase commit index iff
- the majority is greater than the commit index, _and_
- the terms match!

As a side effect, fix a broken test case that committed too much too
early.